refactor: added helper function to upload image list
