+++ src/pages/Dashboard/AddEvent/AddEvent.jsx
@@ -112,6 +112,7 @@ import MultilingualInput from '../../../components/MultilingualInput';
 import { contentLanguageKeyMap } from '../../../constants/contentLanguage';
 import { doesEventExceedNextDay } from '../../../utils/doesEventExceed';
 import SortableTreeSelect from '../../../components/TreeSelectOption/SortableTreeSelect';
+import { uploadImageListHelper } from '../../../utils/uploadImageListHelper';
 
 const { TextArea } = Input;
 
@@ -979,72 +980,6 @@ function AddEvent() {
               ];
             }
 
-            const uploadImageList = async () => {
-              for (let i = 0; i < values.multipleImagesCrop.length; i++) {
-                const file = values.multipleImagesCrop[i]?.originFileObj;
-                if (!file) {
-                  const cropValues = values.multipleImagesCrop[i]?.cropValues || {};
-                  const imageOptions = values.multipleImagesCrop[i]?.imageOptions || {};
-                  if (cropValues)
-                    imageCrop.push({
-                      ...cropValues,
-                      creditText: imageOptions.credit || null,
-                      description: imageOptions.altText || null,
-                      caption: imageOptions.caption || null,
-                    });
-                  else
-                    imageCrop.push({
-                      ...values.multipleImagesCrop[i],
-                      creditText: imageOptions.credit || null,
-                      description: imageOptions.altText || null,
-                      caption: imageOptions.caption || null,
-                    });
-                  continue;
-                }
-
-                const formdata = new FormData();
-                formdata.append('file', file);
-
-                try {
-                  const response = await addImage({ data: formdata, calendarId }).unwrap();
-
-                  // Process each image in the list
-                  const { large, thumbnail } = values.multipleImagesCrop[i]?.cropValues || {};
-                  const { original, height, width } = response?.data || {};
-                  const { altText, credit, caption } = values.multipleImagesCrop[i]?.imageOptions || {};
-
-                  const galleryImage = {
-                    large: {
-                      xCoordinate: large?.x,
-                      yCoordinate: large?.y,
-                      height: large?.height,
-                      width: large?.width,
-                    },
-                    original: {
-                      entityId: original?.entityId ?? null,
-                      height,
-                      width,
-                    },
-                    thumbnail: {
-                      xCoordinate: thumbnail?.x,
-                      yCoordinate: thumbnail?.y,
-                      height: thumbnail?.height,
-                      width: thumbnail?.width,
-                    },
-                    description: altText,
-                    creditText: credit,
-                    caption,
-                  };
-
-                  // Add the processed image to imageCrop
-                  imageCrop.push(galleryImage);
-                } catch (error) {
-                  console.log(error);
-                  throw error; // rethrow to stop further execution
-                }
-              }
-            };
-
             if (values?.dragger?.length > 0 && values?.dragger[0]?.originFileObj) {
               const formdata = new FormData();
               formdata.append('file', values?.dragger[0].originFileObj);
@@ -1086,7 +1021,8 @@ function AddEvent() {
                         },
                       ];
 
-                    if (values.multipleImagesCrop?.length > 0) await uploadImageList();
+                    if (values.multipleImagesCrop?.length > 0)
+                      await uploadImageListHelper(values, addImage, calendarId, imageCrop);
                     eventObj['image'] = imageCrop;
                     addUpdateEventApiHandler(eventObj, toggle, sameAs)
                       .then((id) => resolve(id))
@@ -1101,7 +1037,8 @@ function AddEvent() {
                     element && element[0]?.scrollIntoView({ block: 'center', behavior: 'smooth' });
                   });
             } else {
-              if (values.multipleImagesCrop?.length > 0) await uploadImageList();
+              if (values.multipleImagesCrop?.length > 0)
+                await uploadImageListHelper(values, addImage, calendarId, imageCrop);
               if (
                 values?.draggerWrap &&
                 values?.dragger?.length === 0 &&
