+++ src/pages/Dashboard/CreateNewOrganization/CreateNewOrganization.jsx
@@ -75,6 +75,7 @@ import { getCurrentCalendarDetailsFromUserDetails } from '../../../utils/getCurr
 import { contentLanguageKeyMap } from '../../../constants/contentLanguage';
 import { isDataValid } from '../../../utils/MultiLingualFormItemSupportFunctions';
 import SortableTreeSelect from '../../../components/TreeSelectOption/SortableTreeSelect';
+import { uploadImageListHelper } from '../../../utils/uploadImageListHelper';
 
 function CreateNewOrganization() {
   const timestampRef = useRef(Date.now()).current;
@@ -349,71 +350,6 @@ function CreateNewOrganization() {
               },
             ];
           }
-          const uploadImageList = async () => {
-            for (let i = 0; i < values?.multipleImagesCrop.length; i++) {
-              const file = values.multipleImagesCrop[i]?.originFileObj;
-              if (!file) {
-                const cropValues = values.multipleImagesCrop[i]?.cropValues || {};
-                const imageOptions = values.multipleImagesCrop[i]?.imageOptions || {};
-                if (cropValues)
-                  imageCrop.push({
-                    ...cropValues,
-                    creditText: imageOptions.credit || null,
-                    description: imageOptions.altText || null,
-                    caption: imageOptions.caption || null,
-                  });
-                else
-                  imageCrop.push({
-                    ...values.multipleImagesCrop[i],
-                    creditText: imageOptions.credit || null,
-                    description: imageOptions.altText || null,
-                    caption: imageOptions.caption || null,
-                  });
-                continue;
-              }
-
-              const formdata = new FormData();
-              formdata.append('file', file);
-
-              try {
-                const response = await addImage({ data: formdata, calendarId }).unwrap();
-
-                // Process each image in the list
-                const { large, thumbnail } = values.multipleImagesCrop[i]?.cropValues || {};
-                const { original, height, width } = response?.data || {};
-                const { altText, credit, caption } = values.multipleImagesCrop[i]?.imageOptions || {};
-
-                const galleryImage = {
-                  large: {
-                    xCoordinate: large?.x,
-                    yCoordinate: large?.y,
-                    height: large?.height,
-                    width: large?.width,
-                  },
-                  original: {
-                    entityId: original?.entityId ?? null,
-                    height,
-                    width,
-                  },
-                  thumbnail: {
-                    xCoordinate: thumbnail?.x,
-                    yCoordinate: thumbnail?.y,
-                    height: thumbnail?.height,
-                    width: thumbnail?.width,
-                  },
-                  description: altText,
-                  creditText: credit,
-                  caption,
-                };
-
-                // Add the processed image to imageCrop
-                imageCrop.push(galleryImage);
-              } catch (error) {
-                console.log(error);
-                throw error; // rethrow to stop further execution
-              }
-            }
-          };
 
           if (
             (values?.image || (values?.image && values?.image?.length > 0)) &&
@@ -461,7 +397,8 @@ function CreateNewOrganization() {
                         },
                       ];
 
-                    if (values.multipleImagesCrop?.length > 0) await uploadImageList();
+                    if (values.multipleImagesCrop?.length > 0)
+                      await uploadImageListHelper(values, addImage, calendarId, imageCrop);
                     organizationPayload['image'] = imageCrop;
                     addUpdateOrganizationApiHandler(organizationPayload, toggle)
                       .then((id) => resolve(id))
@@ -476,7 +413,8 @@ function CreateNewOrganization() {
                     element && element[0]?.scrollIntoView({ block: 'center', behavior: 'smooth' });
                   });
             } else {
-              if (values.multipleImagesCrop?.length > 0) await uploadImageList();
+              if (values.multipleImagesCrop?.length > 0)
+                await uploadImageListHelper(values, addImage, calendarId, imageCrop);
               if (
                 values?.image &&
                 values?.image?.length === 0 &&
@@ -510,7 +448,8 @@ function CreateNewOrganization() {
                       large: {},
                       thumbnail: {},
                     };
-                    if (values.multipleImagesCrop?.length > 0) await uploadImageList();
+                    if (values.multipleImagesCrop?.length > 0)
+                      await uploadImageListHelper(values, addImage, calendarId, imageCrop);
                     organizationPayload['image'] = imageCrop;
                     addUpdateOrganizationApiHandler(organizationPayload, toggle);
                   })
@@ -521,7 +460,7 @@ function CreateNewOrganization() {
                   });
             } else {
               if (values.multipleImagesCrop?.length > 0) {
-                await uploadImageList();
+                await uploadImageListHelper(values, addImage, calendarId, imageCrop);
                 organizationPayload['image'] = imageCrop;
               }
               if (values?.logo) {
@@ -580,7 +519,8 @@ function CreateNewOrganization() {
                         },
                       ];
 
-                    if (values.multipleImagesCrop?.length > 0) await uploadImageList();
+                    if (values.multipleImagesCrop?.length > 0)
+                      await uploadImageListHelper(values, addImage, calendarId, imageCrop);
                     organizationPayload['image'] = imageCrop;
                     if (values?.logo?.length > 0 && values?.logo[0]?.originFileObj) {
                       const formdata = new FormData();
@@ -630,7 +570,7 @@ function CreateNewOrganization() {
                   });
             } else {
               if (values.multipleImagesCrop?.length > 0) {
-                await uploadImageList();
+                await uploadImageListHelper(values, addImage, calendarId, imageCrop);
                 organizationPayload['image'] = imageCrop;
               }
               if (
@@ -689,7 +629,7 @@ function CreateNewOrganization() {
               else organizationPayload['logo'] = organizationData?.logo;
             }
             if (values.multipleImagesCrop?.length > 0) {
-              await uploadImageList();
+              await uploadImageListHelper(values, addImage, calendarId, imageCrop);
               organizationPayload['image'] = imageCrop;
             }
             if (
