+++ src/pages/Dashboard/CreateNewPerson/CreateNewPerson.jsx
@@ -55,6 +55,7 @@ import { getCurrentCalendarDetailsFromUserDetails } from '../../../utils/getCurr
 import { contentLanguageKeyMap } from '../../../constants/contentLanguage';
 import { isDataValid } from '../../../utils/MultiLingualFormItemSupportFunctions';
 import SortableTreeSelect from '../../../components/TreeSelectOption/SortableTreeSelect';
+import { uploadImageListHelper } from '../../../utils/uploadImageListHelper';
 
 function CreateNewPerson() {
   const timestampRef = useRef(Date.now()).current;
@@ -287,71 +288,7 @@ function CreateNewPerson() {
             },
           ];
         }
-        const uploadImageList = async () => {
-          for (let i = 0; i < values.multipleImagesCrop.length; i++) {
-            const file = values.multipleImagesCrop[i]?.originFileObj;
-            if (!file) {
-              const cropValues = values.multipleImagesCrop[i]?.cropValues || {};
-              const imageOptions = values.multipleImagesCrop[i]?.imageOptions || {};
-              if (cropValues)
-                imageCrop.push({
-                  ...cropValues,
-                  creditText: imageOptions.credit || null,
-                  description: imageOptions.altText || null,
-                  caption: imageOptions.caption || null,
-                });
-              else
-                imageCrop.push({
-                  ...values.multipleImagesCrop[i],
-                  creditText: imageOptions.credit || null,
-                  description: imageOptions.altText || null,
-                  caption: imageOptions.caption || null,
-                });
-              continue;
-            }
 
-            const formdata = new FormData();
-            formdata.append('file', file);
-
-            try {
-              const response = await addImage({ data: formdata, calendarId }).unwrap();
-
-              // Process each image in the list
-              const { large, thumbnail } = values.multipleImagesCrop[i]?.cropValues || {};
-              const { original, height, width } = response?.data || {};
-              const { altText, credit, caption } = values.multipleImagesCrop[i]?.imageOptions || {};
-
-              const galleryImage = {
-                large: {
-                  xCoordinate: large?.x,
-                  yCoordinate: large?.y,
-                  height: large?.height,
-                  width: large?.width,
-                },
-                original: {
-                  entityId: original?.entityId ?? null,
-                  height,
-                  width,
-                },
-                thumbnail: {
-                  xCoordinate: thumbnail?.x,
-                  yCoordinate: thumbnail?.y,
-                  height: thumbnail?.height,
-                  width: thumbnail?.width,
-                },
-                description: altText,
-                creditText: credit,
-                caption,
-              };
-
-              // Add the processed image to imageCrop
-              imageCrop.push(galleryImage);
-            } catch (error) {
-              console.log(error);
-              throw error; // rethrow to stop further execution
-            }
-          }
-        };
         if (values?.image?.length > 0 && values?.image[0]?.originFileObj) {
           const formdata = new FormData();
           formdata.append('file', values?.image[0].originFileObj);
@@ -393,7 +330,8 @@ function CreateNewPerson() {
                     },
                   ];
 
-                if (values.multipleImagesCrop?.length > 0) await uploadImageList();
+                if (values.multipleImagesCrop?.length > 0)
+                  await uploadImageListHelper(values, addImage, calendarId, imageCrop);
                 personPayload['image'] = imageCrop;
                 addUpdatePersonApiHandler(personPayload);
               })
@@ -403,7 +341,8 @@ function CreateNewPerson() {
                 element && element[0]?.scrollIntoView({ block: 'center', behavior: 'smooth' });
               });
         } else {
-          if (values.multipleImagesCrop?.length > 0) await uploadImageList();
+          if (values.multipleImagesCrop?.length > 0)
+            await uploadImageListHelper(values, addImage, calendarId, imageCrop);
           if (
             values?.image &&
             values?.image?.length === 0 &&
